ndb: remove compiler warnings

+++ sql/ha_ndbcluster.cc
@@ -7337,7 +7337,7 @@ static void print_share(const char* where, NDB_SHARE* share)
   fprintf(DBUG_FILE,
           "%s %s.%s: use_count: %u, commit_count: %llu\n",
           where, share->db, share->table_name, share->use_count,
-          share->commit_count);
+          (long long unsigned int) share->commit_count);
   fprintf(DBUG_FILE,
           "  - key: %s, key_length: %d\n",
           share->key, share->key_length);
